Moved lessonPage and lessonContent to lesson entity
