+++ backend/src/main/java/com/enpm613/algolab/entity/lesson/PracticeQuestion.java
@@ -1,0 +1,14 @@
+package com.enpm613.algolab.entity.lesson;
+
+public class PracticeQuestion {
+
+    private Long practiceQuestionId;
+
+    private String questionDifficulty;
+
+    private String questionContent;
+
+    private String answerContent;
+
+
+}
